Refactor workspace services (#545)
