+++ src/backend/services/workspace-query.service.ts
@@ -1,0 +1,300 @@
+import { type KanbanColumn, WorkspaceStatus } from '@prisma-gen/client';
+import pLimit from 'p-limit';
+import { projectAccessor } from '../resource_accessors/project.accessor';
+import { workspaceAccessor } from '../resource_accessors/workspace.accessor';
+import { FactoryConfigService } from './factory-config.service';
+import { gitOpsService } from './git-ops.service';
+import { githubCLIService } from './github-cli.service';
+import { computeKanbanColumn, kanbanStateService } from './kanban-state.service';
+import { createLogger } from './logger.service';
+import { sessionService } from './session.service';
+
+const logger = createLogger('workspace-query');
+
+// Limit concurrent git operations to prevent resource exhaustion.
+const DEFAULT_GIT_CONCURRENCY = 3;
+const gitConcurrencyLimit = pLimit(DEFAULT_GIT_CONCURRENCY);
+
+// Cache for GitHub review requests (expensive API call)
+let cachedReviewCount: { count: number; fetchedAt: number } | null = null;
+const REVIEW_CACHE_TTL_MS = 60_000; // 1 minute cache
+
+class WorkspaceQueryService {
+  async getProjectSummaryState(projectId: string) {
+    const [project, workspaces] = await Promise.all([
+      projectAccessor.findById(projectId),
+      workspaceAccessor.findByProjectIdWithSessions(projectId, {
+        excludeStatuses: [WorkspaceStatus.ARCHIVED],
+      }),
+    ]);
+
+    const defaultBranch = project?.defaultBranch ?? 'main';
+
+    const workingStatusByWorkspace = new Map<string, boolean>();
+    for (const workspace of workspaces) {
+      const sessionIds = workspace.claudeSessions?.map((s) => s.id) ?? [];
+      workingStatusByWorkspace.set(workspace.id, sessionService.isAnySessionWorking(sessionIds));
+    }
+
+    const gitStatsResults: Record<
+      string,
+      { total: number; additions: number; deletions: number; hasUncommitted: boolean } | null
+    > = {};
+
+    await Promise.all(
+      workspaces.map((workspace) =>
+        gitConcurrencyLimit(async () => {
+          if (!workspace.worktreePath) {
+            gitStatsResults[workspace.id] = null;
+            return;
+          }
+          try {
+            gitStatsResults[workspace.id] = await gitOpsService.getWorkspaceGitStats(
+              workspace.worktreePath,
+              defaultBranch
+            );
+          } catch (error) {
+            logger.debug('Failed to get git stats for workspace', {
+              workspaceId: workspace.id,
+              error: error instanceof Error ? error.message : String(error),
+            });
+            gitStatsResults[workspace.id] = null;
+          }
+        })
+      )
+    );
+
+    let reviewCount = 0;
+    const now = Date.now();
+    if (cachedReviewCount && now - cachedReviewCount.fetchedAt < REVIEW_CACHE_TTL_MS) {
+      reviewCount = cachedReviewCount.count;
+    } else {
+      try {
+        const health = await githubCLIService.checkHealth();
+        if (health.isInstalled && health.isAuthenticated) {
+          const prs = await githubCLIService.listReviewRequests();
+          reviewCount = prs.filter((pr) => pr.reviewDecision !== 'APPROVED').length;
+          cachedReviewCount = { count: reviewCount, fetchedAt: now };
+        }
+      } catch (error) {
+        logger.debug('Failed to fetch review count', {
+          error: error instanceof Error ? error.message : String(error),
+        });
+      }
+    }
+
+    return {
+      workspaces: workspaces.map((w) => {
+        const sessionDates = [
+          ...(w.claudeSessions?.map((s) => s.updatedAt) ?? []),
+          ...(w.terminalSessions?.map((s) => s.updatedAt) ?? []),
+        ].filter(Boolean) as Date[];
+        const lastActivityAt =
+          sessionDates.length > 0
+            ? sessionDates.reduce((latest, d) => (d > latest ? d : latest)).toISOString()
+            : null;
+
+        return {
+          id: w.id,
+          name: w.name,
+          branchName: w.branchName,
+          prUrl: w.prUrl,
+          prNumber: w.prNumber,
+          prState: w.prState,
+          prCiStatus: w.prCiStatus,
+          isWorking: workingStatusByWorkspace.get(w.id) ?? false,
+          gitStats: gitStatsResults[w.id] ?? null,
+          lastActivityAt,
+        };
+      }),
+      reviewCount,
+    };
+  }
+
+  async listWithKanbanState(input: {
+    projectId: string;
+    status?: WorkspaceStatus;
+    kanbanColumn?: KanbanColumn;
+    limit?: number;
+    offset?: number;
+  }) {
+    const { projectId, ...filters } = input;
+
+    const workspaces = await workspaceAccessor.findByProjectIdWithSessions(projectId, {
+      ...filters,
+      excludeStatuses: [WorkspaceStatus.ARCHIVED],
+    });
+
+    return workspaces.map((workspace) => {
+      const sessionIds = workspace.claudeSessions?.map((s) => s.id) ?? [];
+      const isWorking = sessionService.isAnySessionWorking(sessionIds);
+
+      const kanbanColumn = computeKanbanColumn({
+        lifecycle: workspace.status,
+        isWorking,
+        prState: workspace.prState,
+        hasHadSessions: workspace.hasHadSessions,
+      });
+
+      return {
+        ...workspace,
+        kanbanColumn,
+        isWorking,
+      };
+    });
+  }
+
+  async refreshFactoryConfigs(projectId: string) {
+    const workspaces = await workspaceAccessor.findByProjectId(projectId);
+
+    let updatedCount = 0;
+    const errors: Array<{ workspaceId: string; error: string }> = [];
+
+    for (const workspace of workspaces) {
+      if (!workspace.worktreePath) {
+        continue;
+      }
+
+      try {
+        const factoryConfig = await FactoryConfigService.readConfig(workspace.worktreePath);
+
+        await workspaceAccessor.update(workspace.id, {
+          runScriptCommand: factoryConfig?.scripts.run ?? null,
+          runScriptCleanupCommand: factoryConfig?.scripts.cleanup ?? null,
+        });
+
+        updatedCount++;
+      } catch (error) {
+        errors.push({
+          workspaceId: workspace.id,
+          error: error instanceof Error ? error.message : String(error),
+        });
+        logger.error('Failed to refresh factory config for workspace', {
+          workspaceId: workspace.id,
+          error: error instanceof Error ? error.message : String(error),
+        });
+      }
+    }
+
+    return {
+      updatedCount,
+      totalWorkspaces: workspaces.length,
+      errors,
+    };
+  }
+
+  async getFactoryConfig(projectId: string) {
+    const project = await projectAccessor.findById(projectId);
+    if (!project) {
+      throw new Error('Project not found');
+    }
+
+    try {
+      const config = await FactoryConfigService.readConfig(project.repoPath);
+      return config;
+    } catch (error) {
+      logger.error('Failed to read factory config', {
+        projectId,
+        error: error instanceof Error ? error.message : String(error),
+      });
+      return null;
+    }
+  }
+
+  async syncPRStatus(workspaceId: string) {
+    const workspace = await workspaceAccessor.findById(workspaceId);
+    if (!workspace) {
+      throw new Error('Workspace not found');
+    }
+
+    if (!workspace.prUrl) {
+      return { success: false, reason: 'no_pr_url' as const };
+    }
+
+    const prResult = await githubCLIService.fetchAndComputePRState(workspace.prUrl);
+    if (!prResult) {
+      return { success: false, reason: 'fetch_failed' as const };
+    }
+
+    await workspaceAccessor.update(workspaceId, {
+      prNumber: prResult.prNumber,
+      prState: prResult.prState,
+      prReviewState: prResult.prReviewState,
+      prCiStatus: prResult.prCiStatus,
+      prUpdatedAt: new Date(),
+    });
+
+    await kanbanStateService.updateCachedKanbanColumn(workspaceId);
+
+    logger.info('PR status synced manually', {
+      workspaceId,
+      prNumber: prResult.prNumber,
+      prState: prResult.prState,
+    });
+
+    return { success: true, prState: prResult.prState };
+  }
+
+  async syncAllPRStatuses(projectId: string) {
+    const workspaces = await workspaceAccessor.findByProjectIdWithSessions(projectId, {
+      excludeStatuses: [WorkspaceStatus.ARCHIVED],
+    });
+
+    const workspacesWithPRs = workspaces.filter(
+      (w): w is typeof w & { prUrl: string } => w.prUrl !== null
+    );
+
+    if (workspacesWithPRs.length === 0) {
+      return { synced: 0, failed: 0 };
+    }
+
+    let synced = 0;
+    let failed = 0;
+
+    await Promise.all(
+      workspacesWithPRs.map((workspace) =>
+        gitConcurrencyLimit(async () => {
+          const prResult = await githubCLIService.fetchAndComputePRState(workspace.prUrl);
+          if (!prResult) {
+            failed++;
+            return;
+          }
+
+          await workspaceAccessor.update(workspace.id, {
+            prNumber: prResult.prNumber,
+            prState: prResult.prState,
+            prReviewState: prResult.prReviewState,
+            prCiStatus: prResult.prCiStatus,
+            prUpdatedAt: new Date(),
+          });
+
+          await kanbanStateService.updateCachedKanbanColumn(workspace.id);
+          synced++;
+        })
+      )
+    );
+
+    logger.info('Batch PR status sync completed', { projectId, synced, failed });
+
+    return { synced, failed };
+  }
+
+  async hasChanges(workspaceId: string): Promise<boolean> {
+    const workspace = await workspaceAccessor.findByIdWithProject(workspaceId);
+    if (!(workspace?.worktreePath && workspace.project)) {
+      return false;
+    }
+
+    try {
+      const stats = await gitOpsService.getWorkspaceGitStats(
+        workspace.worktreePath,
+        workspace.project.defaultBranch ?? 'main'
+      );
+      return stats !== null && (stats.total > 0 || stats.hasUncommitted);
+    } catch {
+      return false;
+    }
+  }
+}
+
+export const workspaceQueryService = new WorkspaceQueryService();
