+++ src/backend/trpc/workspace.trpc.ts
@@ -1,22 +1,12 @@
-import * as fs from 'node:fs/promises';
-import * as path from 'node:path';
 import { KanbanColumn, WorkspaceStatus } from '@prisma-gen/client';
-import { TRPCError } from '@trpc/server';
-import pLimit from 'p-limit';
 import { z } from 'zod';
-import { GitClientFactory } from '../clients/git.client';
-import { getWorkspaceGitStats } from '../lib/git-helpers';
-import { gitCommand } from '../lib/shell';
-import { projectAccessor } from '../resource_accessors/project.accessor';
 import { workspaceAccessor } from '../resource_accessors/workspace.accessor';
-import { FactoryConfigService } from '../services/factory-config.service';
-import { githubCLIService } from '../services/github-cli.service';
-import { computeKanbanColumn, kanbanStateService } from '../services/kanban-state.service';
 import { createLogger } from '../services/logger.service';
 import { RunScriptService } from '../services/run-script.service';
 import { sessionService } from '../services/session.service';
 import { terminalService } from '../services/terminal.service';
-import { workspaceStateMachine } from '../services/workspace-state-machine.service';
+import { workspaceQueryService } from '../services/workspace-query.service';
+import { worktreeLifecycleService } from '../services/worktree-lifecycle.service';
 import { publicProcedure, router } from './trpc';
 import { workspaceFilesRouter } from './workspace/files.trpc';
 import { workspaceGitRouter } from './workspace/git.trpc';
@@ -30,137 +20,6 @@ export type { GitFileStatus, GitStatusFile } from '../lib/git-helpers';
 export { parseGitStatusOutput } from '../lib/git-helpers';
 
 const logger = createLogger('workspace-trpc');
-
-// Limit concurrent git operations to prevent resource exhaustion.
-// The value 3 is arbitrary but works well in practice - high enough for parallelism,
-// low enough to avoid spawning too many git processes simultaneously.
-const DEFAULT_GIT_CONCURRENCY = 3;
-const gitConcurrencyLimit = pLimit(DEFAULT_GIT_CONCURRENCY);
-
-// Cache for GitHub review requests (expensive API call)
-let cachedReviewCount: { count: number; fetchedAt: number } | null = null;
-const REVIEW_CACHE_TTL_MS = 60_000; // 1 minute cache
-
-type WorkspaceWithProject = Awaited<ReturnType<typeof getWorkspaceWithProjectOrThrow>>;
-
-interface WorktreeCleanupOptions {
-  commitUncommitted: boolean;
-}
-
-function getProjectOrThrow(workspace: WorkspaceWithProject) {
-  const project = workspace.project;
-  if (!(project?.repoPath && project.worktreeBasePath)) {
-    throw new TRPCError({
-      code: 'INTERNAL_SERVER_ERROR',
-      message: 'Workspace project paths are missing',
-    });
-  }
-  return project;
-}
-
-function assertWorktreePathSafe(worktreePath: string, worktreeBasePath: string): void {
-  const resolvedWorktreePath = path.resolve(worktreePath);
-  const resolvedBasePath = path.resolve(worktreeBasePath);
-  const basePrefix = `${resolvedBasePath}${path.sep}`;
-
-  if (resolvedWorktreePath === resolvedBasePath || !resolvedWorktreePath.startsWith(basePrefix)) {
-    throw new TRPCError({
-      code: 'INTERNAL_SERVER_ERROR',
-      message: 'Workspace worktree path is outside the worktree base directory',
-    });
-  }
-}
-
-function pathExists(targetPath: string): Promise<boolean> {
-  return fs
-    .stat(targetPath)
-    .then(() => true)
-    .catch(() => false);
-}
-
-async function commitIfNeeded(
-  worktreePath: string,
-  workspaceName: string,
-  commitUncommitted: boolean
-): Promise<void> {
-  const statusResult = await gitCommand(['status', '--porcelain'], worktreePath);
-  if (statusResult.code !== 0) {
-    throw new TRPCError({
-      code: 'INTERNAL_SERVER_ERROR',
-      message: `Git status failed: ${statusResult.stderr || statusResult.stdout}`,
-    });
-  }
-
-  if (statusResult.stdout.trim().length === 0) {
-    return;
-  }
-
-  if (!commitUncommitted) {
-    throw new TRPCError({
-      code: 'PRECONDITION_FAILED',
-      message: 'Workspace has uncommitted changes. Enable commit-before-archive to proceed.',
-    });
-  }
-
-  const addResult = await gitCommand(['add', '-A'], worktreePath);
-  if (addResult.code !== 0) {
-    throw new TRPCError({
-      code: 'INTERNAL_SERVER_ERROR',
-      message: `Git add failed: ${addResult.stderr || addResult.stdout}`,
-    });
-  }
-
-  const commitMessage = `Archive workspace ${workspaceName}`;
-  const commitResult = await gitCommand(['commit', '-m', commitMessage], worktreePath);
-  if (commitResult.code !== 0) {
-    throw new TRPCError({
-      code: 'INTERNAL_SERVER_ERROR',
-      message: `Git commit failed: ${commitResult.stderr || commitResult.stdout}`,
-    });
-  }
-}
-
-async function removeWorktree(
-  worktreePath: string,
-  project: { repoPath: string; worktreeBasePath: string }
-): Promise<void> {
-  const gitClient = GitClientFactory.forProject({
-    repoPath: project.repoPath,
-    worktreeBasePath: project.worktreeBasePath,
-  });
-  const worktreeName = path.basename(worktreePath);
-
-  const registeredWorktree = await gitClient.checkWorktreeExists(worktreeName);
-  if (registeredWorktree) {
-    await gitClient.deleteWorktree(worktreeName);
-    return;
-  }
-
-  if (await pathExists(worktreePath)) {
-    await fs.rm(worktreePath, { recursive: true, force: true });
-  }
-}
-
-async function cleanupWorkspaceWorktree(
-  workspace: WorkspaceWithProject,
-  options: WorktreeCleanupOptions
-): Promise<void> {
-  const worktreePath = workspace.worktreePath;
-  if (!worktreePath) {
-    return;
-  }
-
-  const project = getProjectOrThrow(workspace);
-  assertWorktreePathSafe(worktreePath, project.worktreeBasePath);
-
-  const worktreeExists = await pathExists(worktreePath);
-  if (!worktreeExists) {
-    return;
-  }
-
-  await commitIfNeeded(worktreePath, workspace.name, options.commitUncommitted);
-  await removeWorktree(worktreePath, project);
-}
 
 // =============================================================================
 // Router
@@ -185,104 +44,7 @@ export const workspaceRouter = router({
   // Get unified project summary state for sidebar (workspaces + working status + git stats + review count)
   getProjectSummaryState: publicProcedure
     .input(z.object({ projectId: z.string() }))
-    .query(async ({ input }) => {
-      // 1. Fetch project (for defaultBranch) and non-archived workspaces with sessions in parallel
-      const [project, workspaces] = await Promise.all([
-        projectAccessor.findById(input.projectId),
-        workspaceAccessor.findByProjectIdWithSessions(input.projectId, {
-          excludeStatuses: [WorkspaceStatus.ARCHIVED],
-        }),
-      ]);
-
-      const defaultBranch = project?.defaultBranch ?? 'main';
-
-      // 2. Compute working status for each workspace
-      const workingStatusByWorkspace = new Map<string, boolean>();
-      for (const workspace of workspaces) {
-        const sessionIds = workspace.claudeSessions?.map((s) => s.id) ?? [];
-        workingStatusByWorkspace.set(workspace.id, sessionService.isAnySessionWorking(sessionIds));
-      }
-
-      // 3. Fetch git stats for all workspaces with concurrency limit
-      const gitStatsResults: Record<
-        string,
-        { total: number; additions: number; deletions: number; hasUncommitted: boolean } | null
-      > = {};
-
-      await Promise.all(
-        workspaces.map((workspace) =>
-          gitConcurrencyLimit(async () => {
-            if (!workspace.worktreePath) {
-              gitStatsResults[workspace.id] = null;
-              return;
-            }
-            try {
-              gitStatsResults[workspace.id] = await getWorkspaceGitStats(
-                workspace.worktreePath,
-                defaultBranch
-              );
-            } catch (error) {
-              logger.debug('Failed to get git stats for workspace', {
-                workspaceId: workspace.id,
-                error: error instanceof Error ? error.message : String(error),
-              });
-              gitStatsResults[workspace.id] = null;
-            }
-          })
-        )
-      );
-
-      // 4. Fetch review count (unapproved PRs only) with caching
-      let reviewCount = 0;
-      const now = Date.now();
-      if (cachedReviewCount && now - cachedReviewCount.fetchedAt < REVIEW_CACHE_TTL_MS) {
-        reviewCount = cachedReviewCount.count;
-      } else {
-        try {
-          const health = await githubCLIService.checkHealth();
-          if (health.isInstalled && health.isAuthenticated) {
-            const prs = await githubCLIService.listReviewRequests();
-            reviewCount = prs.filter((pr) => pr.reviewDecision !== 'APPROVED').length;
-            cachedReviewCount = { count: reviewCount, fetchedAt: now };
-          }
-        } catch (error) {
-          // Review count is non-critical, but log for debugging
-          logger.debug('Failed to fetch review count', {
-            error: error instanceof Error ? error.message : String(error),
-          });
-        }
-      }
-
-      // 5. Build response
-      return {
-        workspaces: workspaces.map((w) => {
-          // Compute last activity from most recent session (Claude or Terminal)
-          // Returns null if no sessions exist (avoid misleading timestamps from metadata changes)
-          const sessionDates = [
-            ...(w.claudeSessions?.map((s) => s.updatedAt) ?? []),
-            ...(w.terminalSessions?.map((s) => s.updatedAt) ?? []),
-          ].filter(Boolean) as Date[];
-          const lastActivityAt =
-            sessionDates.length > 0
-              ? sessionDates.reduce((latest, d) => (d > latest ? d : latest)).toISOString()
-              : null;
-
-          return {
-            id: w.id,
-            name: w.name,
-            branchName: w.branchName,
-            prUrl: w.prUrl,
-            prNumber: w.prNumber,
-            prState: w.prState,
-            prCiStatus: w.prCiStatus,
-            isWorking: workingStatusByWorkspace.get(w.id) ?? false,
-            gitStats: gitStatsResults[w.id] ?? null,
-            lastActivityAt,
-          };
-        }),
-        reviewCount,
-      };
-    }),
+    .query(({ input }) => workspaceQueryService.getProjectSummaryState(input.projectId)),
 
   // List workspaces with kanban state (for board view)
   listWithKanbanState: publicProcedure
@@ -295,37 +57,7 @@ export const workspaceRouter = router({
         offset: z.number().min(0).optional(),
       })
     )
-    .query(async ({ input }) => {
-      const { projectId, ...filters } = input;
-
-      // Get workspaces with sessions included (exclude archived from kanban view at DB level)
-      const workspaces = await workspaceAccessor.findByProjectIdWithSessions(projectId, {
-        ...filters,
-        excludeStatuses: [WorkspaceStatus.ARCHIVED],
-      });
-
-      // Get working status for all workspaces
-      const workspacesWithKanban = workspaces.map((workspace) => {
-        const sessionIds = workspace.claudeSessions?.map((s) => s.id) ?? [];
-        const isWorking = sessionService.isAnySessionWorking(sessionIds);
-
-        // Compute live kanban column
-        const kanbanColumn = computeKanbanColumn({
-          lifecycle: workspace.status,
-          isWorking,
-          prState: workspace.prState,
-          hasHadSessions: workspace.hasHadSessions,
-        });
-
-        return {
-          ...workspace,
-          kanbanColumn,
-          isWorking,
-        };
-      });
-
-      return workspacesWithKanban;
-    }),
+    .query(({ input }) => workspaceQueryService.listWithKanbanState(input)),
 
   // Get workspace by ID
   get: publicProcedure.input(z.object({ id: z.string() })).query(async ({ input }) => {
@@ -394,37 +126,9 @@ export const workspaceRouter = router({
     .input(z.object({ id: z.string(), commitUncommitted: z.boolean().optional() }))
     .mutation(async ({ input }) => {
       const workspace = await getWorkspaceWithProjectOrThrow(input.id);
-      if (!workspaceStateMachine.isValidTransition(workspace.status, 'ARCHIVED')) {
-        throw new TRPCError({
-          code: 'BAD_REQUEST',
-          message: `Cannot archive workspace from status: ${workspace.status}`,
-        });
-      }
-
-      // Clean up running sessions, terminals, and dev processes before archiving
-      try {
-        await sessionService.stopWorkspaceSessions(input.id);
-        await RunScriptService.stopRunScript(input.id);
-        terminalService.destroyWorkspaceTerminals(input.id);
-      } catch (error) {
-        logger.error('Failed to cleanup workspace resources before archive', {
-          workspaceId: input.id,
-          error: error instanceof Error ? error.message : String(error),
-        });
-      }
-
-      try {
-        await cleanupWorkspaceWorktree(workspace, {
-          commitUncommitted: input.commitUncommitted ?? true,
-        });
-      } catch (error) {
-        logger.error('Failed to cleanup workspace worktree before archive', {
-          workspaceId: input.id,
-          error: error instanceof Error ? error.message : String(error),
-        });
-        throw error;
-      }
-      return workspaceStateMachine.archive(input.id);
+      return worktreeLifecycleService.archiveWorkspace(workspace, {
+        commitUncommitted: input.commitUncommitted ?? true,
+      });
     }),
 
   // Delete a workspace
@@ -446,173 +150,27 @@ export const workspaceRouter = router({
   // Refresh factory-factory.json configuration for all workspaces
   refreshFactoryConfigs: publicProcedure
     .input(z.object({ projectId: z.string() }))
-    .mutation(async ({ input }) => {
-      // Get all workspaces for this project that have worktrees
-      const workspaces = await workspaceAccessor.findByProjectId(input.projectId);
-
-      let updatedCount = 0;
-      const errors: Array<{ workspaceId: string; error: string }> = [];
-
-      for (const workspace of workspaces) {
-        if (!workspace.worktreePath) {
-          continue;
-        }
-
-        try {
-          // Read factory-factory.json from the worktree
-          const factoryConfig = await FactoryConfigService.readConfig(workspace.worktreePath);
-
-          // Update workspace with new config
-          await workspaceAccessor.update(workspace.id, {
-            runScriptCommand: factoryConfig?.scripts.run ?? null,
-            runScriptCleanupCommand: factoryConfig?.scripts.cleanup ?? null,
-          });
-
-          updatedCount++;
-        } catch (error) {
-          errors.push({
-            workspaceId: workspace.id,
-            error: error instanceof Error ? error.message : String(error),
-          });
-          logger.error('Failed to refresh factory config for workspace', {
-            workspaceId: workspace.id,
-            error: error instanceof Error ? error.message : String(error),
-          });
-        }
-      }
-
-      return {
-        updatedCount,
-        totalWorkspaces: workspaces.length,
-        errors,
-      };
-    }),
+    .mutation(({ input }) => workspaceQueryService.refreshFactoryConfigs(input.projectId)),
 
   // Get factory-factory.json configuration for a project
   getFactoryConfig: publicProcedure
     .input(z.object({ projectId: z.string() }))
-    .query(async ({ input }) => {
-      const project = await projectAccessor.findById(input.projectId);
-      if (!project) {
-        throw new Error('Project not found');
-      }
-
-      try {
-        const config = await FactoryConfigService.readConfig(project.repoPath);
-        return config;
-      } catch (error) {
-        logger.error('Failed to read factory config', {
-          projectId: input.projectId,
-          error: error instanceof Error ? error.message : String(error),
-        });
-        return null;
-      }
-    }),
+    .query(({ input }) => workspaceQueryService.getFactoryConfig(input.projectId)),
 
   // Sync PR status for a workspace (immediate refresh from GitHub)
   syncPRStatus: publicProcedure
     .input(z.object({ workspaceId: z.string() }))
-    .mutation(async ({ input }) => {
-      const workspace = await workspaceAccessor.findById(input.workspaceId);
-      if (!workspace) {
-        throw new Error('Workspace not found');
-      }
-
-      if (!workspace.prUrl) {
-        return { success: false, reason: 'no_pr_url' as const };
-      }
-
-      const prResult = await githubCLIService.fetchAndComputePRState(workspace.prUrl);
-      if (!prResult) {
-        return { success: false, reason: 'fetch_failed' as const };
-      }
-
-      await workspaceAccessor.update(input.workspaceId, {
-        prNumber: prResult.prNumber,
-        prState: prResult.prState,
-        prReviewState: prResult.prReviewState,
-        prCiStatus: prResult.prCiStatus,
-        prUpdatedAt: new Date(),
-      });
-
-      await kanbanStateService.updateCachedKanbanColumn(input.workspaceId);
-
-      logger.info('PR status synced manually', {
-        workspaceId: input.workspaceId,
-        prNumber: prResult.prNumber,
-        prState: prResult.prState,
-      });
-
-      return { success: true, prState: prResult.prState };
-    }),
+    .mutation(({ input }) => workspaceQueryService.syncPRStatus(input.workspaceId)),
 
   // Sync PR status for all workspaces in a project (immediate refresh from GitHub)
   syncAllPRStatuses: publicProcedure
     .input(z.object({ projectId: z.string() }))
-    .mutation(async ({ input }) => {
-      const workspaces = await workspaceAccessor.findByProjectIdWithSessions(input.projectId, {
-        excludeStatuses: [WorkspaceStatus.ARCHIVED],
-      });
-
-      const workspacesWithPRs = workspaces.filter(
-        (w): w is typeof w & { prUrl: string } => w.prUrl !== null
-      );
-
-      if (workspacesWithPRs.length === 0) {
-        return { synced: 0, failed: 0 };
-      }
-
-      let synced = 0;
-      let failed = 0;
-
-      // Sync all workspaces with PRs (with concurrency limit)
-      await Promise.all(
-        workspacesWithPRs.map((workspace) =>
-          gitConcurrencyLimit(async () => {
-            const prResult = await githubCLIService.fetchAndComputePRState(workspace.prUrl);
-            if (!prResult) {
-              failed++;
-              return;
-            }
-
-            await workspaceAccessor.update(workspace.id, {
-              prNumber: prResult.prNumber,
-              prState: prResult.prState,
-              prReviewState: prResult.prReviewState,
-              prCiStatus: prResult.prCiStatus,
-              prUpdatedAt: new Date(),
-            });
-
-            await kanbanStateService.updateCachedKanbanColumn(workspace.id);
-            synced++;
-          })
-        )
-      );
-
-      logger.info('Batch PR status sync completed', { projectId: input.projectId, synced, failed });
-
-      return { synced, failed };
-    }),
+    .mutation(({ input }) => workspaceQueryService.syncAllPRStatuses(input.projectId)),
 
   // Check if workspace branch has changes relative to the project's default branch
   hasChanges: publicProcedure
     .input(z.object({ workspaceId: z.string() }))
-    .query(async ({ input }) => {
-      const workspace = await workspaceAccessor.findByIdWithProject(input.workspaceId);
-      if (!(workspace?.worktreePath && workspace.project)) {
-        return false;
-      }
-
-      try {
-        const stats = await getWorkspaceGitStats(
-          workspace.worktreePath,
-          workspace.project.defaultBranch ?? 'main'
-        );
-        return stats !== null && (stats.total > 0 || stats.hasUncommitted);
-      } catch {
-        return false;
-      }
-    }),
+    .query(({ input }) => workspaceQueryService.hasChanges(input.workspaceId)),
 
   // Merge sub-routers
   ...workspaceFilesRouter._def.procedures,
